Make gasoline-package a subpackage of funsearch
Management of local packages with pdm is hard for some reason. When running the sandbox, I'd like to call functions from the gasoline-package, but I can't access them for reasons I assume to be related to namespace-issues.
In the sandbox, the funsearch-package is always available (I frankly do not understand why), so by making gasoline a subpackage of funsearch, I can also access gasoline in the sandbox.