Speed up iterative_rounding.py
Because adding constraints to the LP only increases its opt-value (it's minimising), the value only increases over time. Thus, if we match the value of the previous j-iteration's opt-value, we know we also found the opt-value of the current j-iteration's opt-value.

Also, the final_value m.solve() call takes a bit of time (seems to solve an ILP?), and is only necessary at the very end, so I moved it there.