Fix flaw in nemhauser-ullmann implementation
Hopefully for the last time. The error was that `previous_weight` was set on every iteration. If we had multiple identical consecutive items, then all but the first of them would be added to the result.

I tested this implementation against a dumb brute-force implementation, and both agree on the pareto-sets now.